电极选择器界面已重新使用WPF实现
