+++ flashresUniversalSystem/ElectrodeSelecter/BitkyCheckBox.xaml
@@ -1,0 +1,21 @@
+﻿<UserControl x:Class="bitkyFlashresUniversal.ElectrodeSelecter.BitkyCheckBox"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:bitkyFlashresUniversal.ElectrodeSelecter"
+             mc:Ignorable="d">
+    <Grid>
+        <CheckBox x:Name="checkBox" Content="-1" Margin="30,5" Foreground="White" Background="White" RenderTransformOrigin="0.5,0.5" Padding="0" VerticalContentAlignment="Center">
+            <CheckBox.RenderTransform>
+                <TransformGroup>
+                    <ScaleTransform ScaleX="1.5" ScaleY="1.5"/>
+                    <SkewTransform/>
+                    <RotateTransform/>
+                    <TranslateTransform/>
+                </TransformGroup>
+            </CheckBox.RenderTransform>
+        </CheckBox>
+
+    </Grid>
+</UserControl>
