+++ flashresUniversalSystem/ElectrodeSelecter/BitkyCheckBox.xaml.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace bitkyFlashresUniversal.ElectrodeSelecter
+{
+    /// <summary>
+    /// BitkyCheckBox.xaml 的交互逻辑
+    /// </summary>
+    public partial class BitkyCheckBox : UserControl
+    {
+        public BitkyCheckBox()
+        {
+            InitializeComponent();
+        }
+    }
+}
