+++ flashresUniversalSystem/dataExport/DataExport.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using bitkyFlashresUniversal.dataExport.bean;
