+++ flashresUniversalSystem/flashresUniversalSystem.csproj
@@ -72,9 +72,6 @@
     <ApplicationManifest>Properties\app.manifest</ApplicationManifest>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="CSkin">
-      <HintPath>..\..\..\import\CSkin4\CSkin.dll</HintPath>
-    </Reference>
     <Reference Include="Newtonsoft.Json, Version=9.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
       <Private>True</Private>
@@ -85,8 +82,6 @@
       <HintPath>..\packages\System.Data.SQLite.Core.1.0.103\lib\net46\System.Data.SQLite.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xaml">
@@ -167,12 +162,6 @@
     <Compile Include="databaseUtil\SqliteBitky.cs" />
     <Compile Include="ElectrodeDetection\ElectrodeInspect.cs" />
     <Compile Include="ElectrodeDetection\PoleDetectPresenter.cs" />
-    <Compile Include="ElectrodeSelecter\ElectrodeSelecterForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="ElectrodeSelecter\ElectrodeSelecterForm.Designer.cs">
-      <DependentUpon>ElectrodeSelecterForm.cs</DependentUpon>
-    </Compile>
     <Compile Include="ElectrodeSelecter\ProcessPresenter.cs" />
     <Compile Include="poleInfoShow\BitkyPoleControl.xaml.cs">
       <DependentUpon>BitkyPoleControl.xaml</DependentUpon>
@@ -190,9 +179,6 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <EmbeddedResource Include="ElectrodeSelecter\ElectrodeSelecterForm.resx">
-      <DependentUpon>ElectrodeSelecterForm.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
